added authToken checking implementation

+++ server/src/main/java/server/websocket/WebSocketHandler.java
@@ -10,6 +10,7 @@ import model.GameData;
 import org.eclipse.jetty.websocket.api.Session;
 import org.eclipse.jetty.websocket.api.annotations.OnWebSocketMessage;
 import org.eclipse.jetty.websocket.api.annotations.WebSocket;
+import responses.ListGamesResponse;
 import server.Server;
 import websocket.commands.*;
 import websocket.messages.ServerError;
@@ -17,12 +18,14 @@ import websocket.messages.ServerLoadGame;
 import websocket.messages.ServerNotification;
 
 import java.io.IOException;
+import java.util.List;
 
 
 @WebSocket
 public class WebSocketHandler {
 
     private final ConnectionManager connections = new ConnectionManager();
+    private final DataBaseAuthDataAccessObject dataBaseAuthDataAccessObject = new DataBaseAuthDataAccessObject();
 
     @OnWebSocketMessage
     public void onMessage(Session session, String message) {
@@ -56,31 +59,56 @@ public class WebSocketHandler {
 
 
     private void connect(String username, Session session,ConnectCommand connectCommand) throws IOException {
-        connections.add(username, session, connectCommand.getGameID());
-        String color;
-        if (connectCommand.getColor() == ChessGame.TeamColor.WHITE){
-            color = "white";
+        try {
+            if (dataBaseAuthDataAccessObject.getAuth(connectCommand.getAuthString())) {
+                if (new DataBaseGameDataAccessObject().gameExists(connectCommand.getGameID())){
+                connections.add(username, session, connectCommand.getGameID());
+
+                String color;
+                if (connectCommand.getColor() == ChessGame.TeamColor.WHITE) {
+                    color = "white";
+                } else if (connectCommand.getColor() == ChessGame.TeamColor.BLACK) {
+                    color = "black";
+                } else {
+                    color = "observer";
+                }
+                var message = String.format("        %s has joined the game as %s", username, color);
+                ServerNotification notification = new ServerNotification(message);
+                ServerLoadGame serverLoadGame = new ServerLoadGame(true, connectCommand.getGameID(), connectCommand.getColor());
+                connections.send(username, serverLoadGame, connectCommand.getGameID());
+                connections.broadcast(username, notification, connectCommand.getGameID());
+            }
+                else{
+                    throw new DataAccessException("        bad game ID");
+                }
+            }
+            else{
+                throw new DataAccessException("        unauthorized");
+            }
         }
-        else if (connectCommand.getColor() == ChessGame.TeamColor.BLACK){
-            color = "black";
+        catch (DataAccessException e){
+            var serverError = new ServerError(e.getMessage());
+            session.getRemote().sendString(new Gson().toJson(serverError));
         }
-        else{
-            color = "observer";
-        }
-        var message = String.format("        %s has joined the game as %s", username,color);
-        ServerNotification notification = new ServerNotification(message);
-        ServerLoadGame serverLoadGame = new ServerLoadGame(true,connectCommand.getGameID(),connectCommand.getColor());
-        connections.send(username,serverLoadGame, connectCommand.getGameID());
-        connections.broadcast(username, notification, connectCommand.getGameID());
     }
 
     private void makeMove(String username,Session session, MakeMoveCommand makeMoveCommand) throws IOException {
         try {
-            GameData gameData = new DataBaseGameDataAccessObject().getGame(makeMoveCommand.getGameID());
-            ChessGame updatedGame = gameData.game();
-            updatedGame.makeMove(makeMoveCommand.getChessMove());
-            GameData updatedGameData = new GameData(gameData.gameID(),gameData.whiteUsername(),gameData.blackUsername(),gameData.gameName(),updatedGame);
-            new DataBaseGameDataAccessObject().updateGame(makeMoveCommand.getGameID(), updatedGameData);
+            if (dataBaseAuthDataAccessObject.getAuth(makeMoveCommand.getAuthString())) {
+                GameData gameData = new DataBaseGameDataAccessObject().getGame(makeMoveCommand.getGameID());
+                ChessGame updatedGame = gameData.game();
+                updatedGame.makeMove(makeMoveCommand.getChessMove());
+                GameData updatedGameData = new GameData(gameData.gameID(), gameData.whiteUsername(), gameData.blackUsername(), gameData.gameName(), updatedGame);
+                new DataBaseGameDataAccessObject().updateGame(makeMoveCommand.getGameID(), updatedGameData);
+            }
+            else{
+                throw new DataAccessException("        unauthorized");
+            }
+            }
+        catch (DataAccessException e){
+            var serverError = new ServerError(e.getMessage());
+            session.getRemote().sendString(new Gson().toJson(serverError));
+            return;
         }
         catch (InvalidMoveException e){
             ServerError serverError = new ServerError(e.getMessage());
@@ -101,21 +129,37 @@ public class WebSocketHandler {
 
     private void leaveGame(Session session, String username, LeaveGameCommand leaveGameCommand) throws IOException {
         try {
-            connections.remove(leaveGameCommand.getGameID(), username, leaveGameCommand.getColor());
-            var message = String.format("        %s has left the game", username);
-            var notification = new ServerNotification(message);
-            connections.broadcast(username, notification, leaveGameCommand.getGameID());
+            if (dataBaseAuthDataAccessObject.getAuth(leaveGameCommand.getAuthString())) {
+                connections.remove(leaveGameCommand.getGameID(), username, leaveGameCommand.getColor());
+                var message = String.format("        %s has left the game", username);
+                var notification = new ServerNotification(message);
+                connections.broadcast(username, notification, leaveGameCommand.getGameID());
+            }
+            else{
+                throw new DataAccessException("        unauthorized");
+            }
         }
         catch (DataAccessException e){
-            System.out.println(e.getMessage());
+            var serverError = new ServerError(e.getMessage());
+            session.getRemote().sendString(new Gson().toJson(serverError));
         }
     }
 
     private void resign(Session session, String username, ResignCommand resignCommand) throws IOException {
-        var message = String.format("        %s has forfeit", username);
-        var notification = new ServerNotification(message);
-        connections.broadcast(username,notification, resignCommand.getGameID());
-        connections.deleteGame(resignCommand.getGameID());
+        try {
+            if (dataBaseAuthDataAccessObject.getAuth(resignCommand.getAuthString())) {
+                var message = String.format("        %s has forfeit", username);
+                var notification = new ServerNotification(message);
+                connections.broadcast(username, notification, resignCommand.getGameID());
+                connections.deleteGame(resignCommand.getGameID());
+            } else {
+                throw new DataAccessException("        unauthorized");
+            }
+        }
+        catch (DataAccessException e){
+            var serverError = new ServerError(e.getMessage());
+            session.getRemote().sendString(new Gson().toJson(serverError));
+        }
     }
 
 }
